+++ doc/src/modules/printing.rst
@@ -580,8 +580,6 @@ This class is a base class for other classes that implement code-printing
 functionality, and additionally lists a number of functions that cannot be
 easily translated to C or Fortran.
 
-.. autoclass:: sympy.printing.codeprinter.Assignment
-
 .. autoclass:: sympy.printing.codeprinter.CodePrinter
 
    .. autoattribute:: CodePrinter.printmethod
