+++ sympy/__init__.py
@@ -516,7 +516,6 @@ __all__.extend([
     'algebras',
     'assumptions',
     'calculus',
-    'codegen',
     'combinatorics',
     'concrete',
     'deprecated',
